+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/network/node/NetworkNodeDestructor.java
@@ -26,6 +26,8 @@ import net.minecraft.tileentity.TileEntityShulkerBox;
 import net.minecraft.util.NonNullList;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.RayTraceResult;
+import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
 import net.minecraft.world.WorldServer;
 import net.minecraft.world.chunk.Chunk;
@@ -102,7 +104,13 @@ public class NetworkNodeDestructor extends NetworkNode implements IComparable, I
                 IBlockState frontBlockState = world.getBlockState(front);
                 Block frontBlock = frontBlockState.getBlock();
 
-                ItemStack frontStack = frontBlock.getPickBlock(frontBlockState, null, world, front, null);
+                ItemStack frontStack = frontBlock.getPickBlock(
+                    frontBlockState,
+                    new RayTraceResult(new Vec3d(pos.getX(), pos.getY(), pos.getZ()), getDirection().getOpposite()),
+                    world,
+                    front,
+                    FakePlayerFactory.getMinecraft((WorldServer) world)
+                );
 
                 if (!frontStack.isEmpty()) {
                     if (IFilterable.canTake(itemFilters, mode, compare, frontStack) && frontBlockState.getBlockHardness(world, front) != -1.0) {