+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - You can no longer put a Filter in filter slots to gain additional filter slots (raoulvdberge)
 - You can now re-insert Processing Patterns in the Pattern Grid and have the inputs and outputs be completed (raoulvdberge)
 - Fixed bug where pattern was recipe pattern was creatable when there was no recipe output (raoulvdberge)
+- Fixed a crash when breaking an Ender IO conduit with the Destructor (raoulvdberge)
 - Added a missing config option for Crafter Manager energy usage (raoulvdberge)
 - If an Interface is configured to expose the entire network storage (by configuring no export slots), it will no longer expose the entire RS storage, due to performance issues (raoulvdberge)
 - The Portable Grid no longer exposes a inventory for crossmod interaction, due to performance issues (raoulvdberge)